tests: move long files to separate directories
