+++ lumicks/pylake/tests/test_imaging_confocal/test_point_scan.py
