+++ lumicks/pylake/tests/test_imaging_confocal/test_scan.py
